refactor : survey result structure
