+++ components/layouts/index.ts
@@ -1,2 +1,3 @@
 export * from './survey-home';
 export * from './survey-processing';
+export * from './survey-result';
