--- components/modules/survey-result/components/survey-result-explain/survey-result-explain.component.tsx
@@ -1,11 +1,0 @@
-import { Typography } from 'antd';
-import { FC } from 'react';
-type ISurveyResultExplainProp = {};
-
-export const SurveyResultExplain: FC<ISurveyResultExplainProp> = () => {
-  return (
-    <>
-        <Typography>짱구!!!</Typography>
-    </>
-  );
-};
