--- components/modules/survey-result/index.ts
@@ -1,2 +1,0 @@
-export * from './components';
-export * from './survey-result-module'
