+++ pages/survey/result/index.tsx
@@ -1,7 +1,6 @@
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
-import { SurveyResultScreen } from '../../screens/survey/survey-result';
-
+import { SurveyResultScreen } from '../../../screens';
 export const getServerSideProps: GetServerSideProps = async (context) => {
   return {
     props: {},
@@ -12,7 +11,7 @@ const SurveyResultPage = () => {
   return (
     <>
       <Head>
-        <title>설문지 | 네잎클로바</title>
+        <title>검사 결과 | 네잎클로바</title>
       </Head>
       <SurveyResultScreen />
     </>
