+++ screens/survey/survey-result/components/index.ts
