+++ screens/survey/survey-result/components/survey-result-explain/index.ts
