+++ screens/survey/survey-result/components/survey-result-explain/survey-result-explain.component.tsx
@@ -1,0 +1,11 @@
+import { Typography } from 'antd';
+import { FC } from 'react';
+type ISurveyResultExplainComponentProp = {};
+
+export const SurveyResultExplainComponent: FC<ISurveyResultExplainComponentProp> = () => {
+  return (
+    <>
+      <Typography>짱구!!!</Typography>
+    </>
+  );
+};
