+++ screens/survey/survey-result/components/survey-result-title/images/index.ts
