+++ screens/survey/survey-result/components/survey-result-title/index.ts
