+++ screens/survey/survey-result/components/survey-result-title/survey-result-title.component.tsx
@@ -1,0 +1,14 @@
+import { Typography } from 'antd';
+import Image from 'next/image';
+import { FC } from 'react';
+import zzangoo from './images/zzangoo.png';
+type ISurveyResultTitleComponentProp = {};
+
+export const SurveyResultTitleComponent: FC<ISurveyResultTitleComponentProp> = () => {
+  return (
+    <>
+      <Typography>당신은 바로~~~</Typography>
+      <Image src={zzangoo} alt="짱구그림" />
+    </>
+  );
+};
