+++ screens/survey/survey-result/survey-result.screen.tsx
@@ -1,13 +1,19 @@
+import { Space } from 'antd';
 import { NextPage } from 'next';
-import { SurveyResultModule } from '../../../components';
-import { SurveyResultLayout } from '../../../components/layouts/survey-result';
+import { SurveyResultLayout } from '../../../components';
+import { SurveyResultExplainComponent, SurveyResultTitleComponent } from './components';
 
 type ISurveyResultScreenProps = {};
 
 export const SurveyResultScreen: NextPage<ISurveyResultScreenProps> = () => {
   return (
     <SurveyResultLayout>
-      <SurveyResultModule />
+      <div style={{ display: 'flex', padding: 20, justifyContent: 'center' }}>
+        <Space direction="vertical">
+          <SurveyResultTitleComponent />
+          <SurveyResultExplainComponent />
+        </Space>
+      </div>
     </SurveyResultLayout>
   );
 };
